refactor(web): move MIME table to http-helpers where it belongs
The MIME extension→content-type lookup was sitting in atomic-write.ts,
but it has nothing to do with atomic file writes — it's only used by
serveFile in http-helpers.ts. Moves the constant to its actual caller
and drops the cross-module import.

No behavior change.
